new_beginnings/requirements.txt
django
djangorestframework
django-phonenumber-field
django-phonenumbers